Don't restart workspace in case of errors
